Fixex Location in CLI. Started model methods.
