+++ bin/run.rb
@@ -1,5 +1,5 @@
 require_relative '../config/environment'
 
 
-Ascii.title
-run_program
+# Ascii.title
+# run_program
