+++ config/environment.rb
@@ -3,5 +3,5 @@ Bundler.require
 
 ActiveRecord::Base.establish_connection(adapter: 'sqlite3', database: 'db/development.db')
 require_all 'lib'
-# require_all 'bin'
+require_all 'bin'
