+++ db/seeds.rb
@@ -8,7 +8,7 @@ href_arr.each do |state_href|
     href = "#{state_href}"
     haunt_hasher("#{state_href}").each do |one, two|
         two.each do |hash|
-            Haunt.find_or_create_by({ name: "#{hash[:name]}", city: "#{hash[:city]}", state: "#{hash[:state]}", description: "#{hash[:description]}"}).save
+            Haunt.find_or_create_by({ name: "#{hash[:name]}", city: "#{hash[:city]}", state: "#{hash[:state]}", description: "#{hash[:description]}"})
         end
     end
 end
@@ -17,7 +17,7 @@ end
 Haunt.all.each do |haunt|
     arr = []
     activities.each do |key, value|
-        desc = haunt.descriptione
+        desc = haunt.description
         descrip = desc.split(/\W+/)
         matches = descrip & value
         if matches.length > 0
@@ -25,7 +25,7 @@ Haunt.all.each do |haunt|
         end
     end
     exp = arr.uniq.join(" ")
-    Paranormal_Experience.create(experience: "#{exp}")
+    ParanormalExperience.create(experience: "#{exp}")
 end
