+++ lib/tester.rb
@@ -1,0 +1,9 @@
+require_relative '../config/environment'
+
+# x = ParanormalExperience.find_by_id(123)
+
+# puts x.haunt_by_experience
+
+y = Haunt.last
+
+puts y.location_of_haunt
